Refactor: Remove unused CSS styles, delete test HTML file, and implement card and search settings management modules
- Deleted `style.css` and `test-filter.html` files as they were no longer needed.
- Updated `style.css` to remove the `.locale-container` class.
- Added `cardManager.js` for managing bookmarks and history cards, including loading, rendering, and configuration.
- Introduced `searchSettings.js` for managing search bar configurations and rendering.
- Implemented local storage functionality for saving and loading configurations.
- Enhanced search functionality with support for multiple search engines.
