+++ src/cardManager.js
@@ -1,0 +1,546 @@
+// ====== 卡片管理模块 ======
+// 负责卡片的数据加载、渲染、配置管理和布局设置
+
+import { FilterParser, filterItems } from './filterParser.js';
+
+// ====== 常量 ======
+const BOOKMARK_ROOT_SENTINEL = '__BOOKMARK_ROOT__';
+
+// ====== 状态 ======
+let allBookmarks = [];
+let allHistory = [];
+let cardsConfig = [];
+let layoutConfig = { cardsPerRow: 2, cardHeight: 250 };
+let currentEditingCardId = null;
+let openInNewTab = true; // 从 searchSettings 导入
+
+// ====== 本地存储 ======
+function loadCardsConfig() {
+    const data = localStorage.getItem("cardsConfig");
+    if (data) {
+        return JSON.parse(data);
+    }
+    return [];
+}
+
+function saveCardsConfig(config) {
+    localStorage.setItem("cardsConfig", JSON.stringify(config));
+}
+
+function loadLayoutConfig() {
+    const data = localStorage.getItem("layoutConfig");
+    if (data) {
+        return JSON.parse(data);
+    }
+    return { cardsPerRow: 2, cardHeight: 250 };
+}
+
+function saveLayoutConfig(config) {
+    localStorage.setItem("layoutConfig", JSON.stringify(config));
+}
+
+function generateCardId() {
+    return 'card_' + Date.now() + '_' + Math.random().toString(36).substr(2, 9);
+}
+
+// ====== 数据加载功能 ======
+async function loadAllData() {
+    await Promise.all([loadBookmarks(), loadHistory()]);
+    renderAllCards();
+}
+
+async function loadBookmarks() {
+    try {
+        if (typeof chrome !== 'undefined' && chrome.bookmarks) {
+            const bookmarks = await chrome.bookmarks.getTree();
+            allBookmarks = extractBookmarks(bookmarks);
+        } else {
+            console.warn('Chrome bookmarks API not available');
+        }
+    } catch (error) {
+        console.error('Error loading bookmarks:', error);
+    }
+}
+
+async function loadHistory() {
+    try {
+        if (typeof chrome !== 'undefined' && chrome.history) {
+            const historyItems = await chrome.history.search({
+                text: '',
+                startTime: 0,
+                maxResults: 2000
+            });
+            
+            allHistory = historyItems.map(item => ({
+                title: item.title || item.url,
+                url: item.url,
+                visitCount: item.visitCount,
+                lastVisitTime: new Date(item.lastVisitTime),
+                typedCount: item.typedCount
+            }));
+            
+            // 按最后访问时间降序排序
+            allHistory.sort((a, b) => b.lastVisitTime - a.lastVisitTime);
+        } else {
+            console.warn('Chrome history API not available');
+        }
+    } catch (error) {
+        console.error('Error loading history:', error);
+    }
+}
+
+function extractBookmarks(bookmarkNodes, folderPath = '') {
+    let result = [];
+    
+    for (const node of bookmarkNodes) {
+        if (node.children) {
+            // 这是一个文件夹
+            const currentPath = folderPath ? `${folderPath} > ${node.title}` : node.title;
+            result = result.concat(extractBookmarks(node.children, currentPath));
+        } else if (node.url) {
+            // 这是一个书签
+            result.push({
+                title: node.title,
+                url: node.url,
+                folder: folderPath || BOOKMARK_ROOT_SENTINEL
+            });
+        }
+    }
+    
+    return result;
+}
+
+// ====== 卡片渲染功能 ======
+function renderAllCards() {
+    const container = document.getElementById('cards-container');
+    container.innerHTML = '';
+    
+    if (cardsConfig.length === 0) {
+        container.innerHTML = `<div class="empty-message">暂无卡片，点击"添加卡片"开始创建</div>`;
+        return;
+    }
+    
+    // 创建卡片网格
+    const cardsGrid = document.createElement('div');
+    cardsGrid.className = 'cards-grid';
+    cardsGrid.style.gridTemplateColumns = `repeat(${layoutConfig.cardsPerRow}, 1fr)`;
+    
+    cardsConfig.forEach(cardConfig => {
+        const cardElement = createCardElement(cardConfig);
+        cardsGrid.appendChild(cardElement);
+    });
+    
+    container.appendChild(cardsGrid);
+}
+
+function createCardElement(cardConfig) {
+    const card = document.createElement('div');
+    card.className = 'content-card';
+    card.style.height = `${layoutConfig.cardHeight}px`;
+    
+    // 卡片头部
+    const cardHeader = document.createElement('div');
+    cardHeader.className = 'card-header';
+    
+    const cardTitle = document.createElement('h4');
+    cardTitle.className = 'card-title';
+    cardTitle.textContent = cardConfig.title;
+    
+    const cardActions = document.createElement('div');
+    cardActions.className = 'card-actions';
+    
+    const duplicateButton = document.createElement('button');
+    duplicateButton.className = 'card-action-button';
+    duplicateButton.innerHTML = '📋';
+    duplicateButton.title = '复制卡片';
+    duplicateButton.addEventListener('click', () => duplicateCard(cardConfig.id));
+    
+    const editButton = document.createElement('button');
+    editButton.className = 'card-action-button';
+    editButton.innerHTML = '⚙️';
+    editButton.title = '编辑卡片';
+    editButton.addEventListener('click', () => editCard(cardConfig.id));
+    
+    const deleteButton = document.createElement('button');
+    deleteButton.className = 'card-action-button';
+    deleteButton.innerHTML = '🗑️';
+    deleteButton.title = '删除卡片';
+    deleteButton.addEventListener('click', () => deleteCard(cardConfig.id));
+    
+    cardActions.appendChild(duplicateButton);
+    cardActions.appendChild(editButton);
+    cardActions.appendChild(deleteButton);
+    
+    cardHeader.appendChild(cardTitle);
+    cardHeader.appendChild(cardActions);
+    
+    // 卡片内容
+    const cardContent = document.createElement('div');
+    cardContent.className = 'card-content';
+    
+    renderCardContent(cardContent, cardConfig);
+    
+    card.appendChild(cardHeader);
+    card.appendChild(cardContent);
+    
+    return card;
+}
+
+function renderCardContent(container, cardConfig) {
+    const items = getFilteredItems(cardConfig);
+    
+    if (items.length === 0) {
+        container.innerHTML = `<div class="card-empty">暂无内容</div>`;
+        return;
+    }
+    
+    const itemsList = document.createElement('div');
+    itemsList.className = 'card-items-list';
+    
+    items.slice(0, cardConfig.maxItems).forEach(item => {
+        const itemElement = createItemElement(item);
+        itemsList.appendChild(itemElement);
+    });
+    
+    container.innerHTML = '';
+    container.appendChild(itemsList);
+}
+
+function createItemElement(item) {
+    const itemDiv = document.createElement('div');
+    itemDiv.className = 'card-item';
+    
+    const favicon = document.createElement('img');
+    favicon.className = 'card-item-favicon';
+    
+    try {
+        const hostname = new URL(item.url).hostname;
+        favicon.src = `https://www.google.com/s2/favicons?domain=${hostname}&sz=32`;
+    } catch (e) {
+        favicon.style.display = 'none';
+    }
+    
+    favicon.onerror = () => {
+        favicon.style.display = 'none';
+    };
+    
+    const itemInfo = document.createElement('div');
+    itemInfo.className = 'card-item-info';
+    
+    const title = document.createElement('div');
+    title.className = 'card-item-title';
+    title.textContent = item.title || item.url;
+    title.title = item.title || item.url;
+    
+    const meta = document.createElement('div');
+    meta.className = 'card-item-meta';
+    
+    // 根据数据源类型显示不同的元信息
+    if (item.folder) {
+        // 来自收藏夹
+        const folderLabel = item.folder === BOOKMARK_ROOT_SENTINEL ? '根目录' : item.folder;
+        meta.textContent = `📑 ${folderLabel}`;
+    } else if (item.lastVisitTime) {
+        // 来自历史记录
+        const timeStr = formatTimeAgo(item.lastVisitTime);
+        meta.textContent = `🕒 访问${item.visitCount}次 · ${timeStr}`;
+    }
+    
+    itemInfo.appendChild(title);
+    itemInfo.appendChild(meta);
+    
+    itemDiv.appendChild(favicon);
+    itemDiv.appendChild(itemInfo);
+    
+    // 点击事件
+    itemDiv.addEventListener('click', () => {
+        if (openInNewTab) {
+            window.open(item.url, '_blank');
+        } else {
+            window.open(item.url, '_self');
+        }
+    });
+    
+    return itemDiv;
+}
+
+function getFilteredItems(cardConfig) {
+    let items = [];
+    
+    // 支持旧版本单一 type 字段
+    const types = cardConfig.types || (cardConfig.type ? [cardConfig.type] : []);
+    
+    // 从选中的数据源收集数据
+    if (types.includes('bookmarks')) {
+        let bookmarkItems = [...allBookmarks];
+        
+        // 文件夹筛选
+        if (cardConfig.folderFilter) {
+            const folderFilter = cardConfig.folderFilter.toLowerCase();
+            bookmarkItems = bookmarkItems.filter(item => 
+                item.folder && item.folder.toLowerCase().includes(folderFilter)
+            );
+        }
+        
+        items = items.concat(bookmarkItems);
+    }
+    
+    if (types.includes('history')) {
+        let historyItems = [...allHistory];
+        
+        // 时间筛选
+        if (cardConfig.historyDays && cardConfig.historyDays > 0) {
+            const cutoffTime = Date.now() - (cardConfig.historyDays * 24 * 60 * 60 * 1000);
+            historyItems = historyItems.filter(item => item.lastVisitTime.getTime() >= cutoffTime);
+        }
+        
+        items = items.concat(historyItems);
+    }
+    
+    // 去重（同一个 URL 可能既在收藏夹又在历史记录中）
+    const uniqueItems = [];
+    const seenUrls = new Set();
+    
+    for (const item of items) {
+        if (!seenUrls.has(item.url)) {
+            seenUrls.add(item.url);
+            uniqueItems.push(item);
+        }
+    }
+    
+    items = uniqueItems;
+    
+    // 关键词筛选 - 使用筛选解析器
+    if (cardConfig.filter) {
+        items = filterItems(cardConfig.filter, items);
+    }
+    
+    return items;
+}
+
+function formatTimeAgo(date) {
+    const now = new Date();
+    const diffMs = now - date;
+    const diffMins = Math.floor(diffMs / 60000);
+    const diffHours = Math.floor(diffMins / 60);
+    const diffDays = Math.floor(diffHours / 24);
+    
+    if (diffMins < 1) return '刚刚';
+    if (diffMins < 60) return `${diffMins}分钟前`;
+    if (diffHours < 24) return `${diffHours}小时前`;
+    if (diffDays < 7) return `${diffDays}天前`;
+    
+    return date.toLocaleDateString('zh-CN');
+}
+
+// ====== 卡片管理功能 ======
+function addCard() {
+    currentEditingCardId = null;
+    clearCardConfigForm();
+    showCardConfigModal('添加卡片');
+}
+
+function duplicateCard(cardId) {
+    const cardConfig = cardsConfig.find(c => c.id === cardId);
+    if (!cardConfig) return;
+    
+    // 创建副本
+    const duplicatedCard = {
+        ...cardConfig,
+        id: generateCardId(),
+        title: `${cardConfig.title}（副本）`
+    };
+    
+    cardsConfig.push(duplicatedCard);
+    saveCardsConfig(cardsConfig);
+    renderAllCards();
+}
+
+function editCard(cardId) {
+    currentEditingCardId = cardId;
+    const cardConfig = cardsConfig.find(c => c.id === cardId);
+    if (!cardConfig) return;
+    
+    fillCardConfigForm(cardConfig);
+    showCardConfigModal('编辑卡片');
+}
+
+function deleteCard(cardId) {
+    if (confirm('确定要删除这个卡片吗？')) {
+        cardsConfig = cardsConfig.filter(c => c.id !== cardId);
+        saveCardsConfig(cardsConfig);
+        renderAllCards();
+    }
+}
+
+function showCardConfigModal(title) {
+    document.getElementById('modal-title').textContent = title;
+    document.getElementById('card-config-modal').style.display = 'block';
+}
+
+function hideCardConfigModal() {
+    document.getElementById('card-config-modal').style.display = 'none';
+}
+
+function clearCardConfigForm() {
+    document.getElementById('card-title').value = '';
+    document.getElementById('card-type-bookmarks').checked = false;
+    document.getElementById('card-type-history').checked = false;
+    document.getElementById('card-filter').value = '';
+    document.getElementById('card-folder-filter').value = '';
+    document.getElementById('card-history-days').value = '7';
+    document.getElementById('card-max-items').value = '10';
+    toggleCardTypeFields();
+}
+
+function fillCardConfigForm(cardConfig) {
+    document.getElementById('card-title').value = cardConfig.title || '';
+    
+    // 支持旧版本单一 type 字段
+    const types = cardConfig.types || (cardConfig.type ? [cardConfig.type] : []);
+    document.getElementById('card-type-bookmarks').checked = types.includes('bookmarks');
+    document.getElementById('card-type-history').checked = types.includes('history');
+    
+    document.getElementById('card-filter').value = cardConfig.filter || '';
+    document.getElementById('card-folder-filter').value = cardConfig.folderFilter || '';
+    document.getElementById('card-history-days').value = cardConfig.historyDays || '7';
+    document.getElementById('card-max-items').value = cardConfig.maxItems || '10';
+    toggleCardTypeFields();
+}
+
+function toggleCardTypeFields() {
+    const bookmarksChecked = document.getElementById('card-type-bookmarks').checked;
+    const historyChecked = document.getElementById('card-type-history').checked;
+    const historyField = document.getElementById('history-time-field');
+    const bookmarksField = document.getElementById('bookmarks-folder-field');
+    
+    // 显示/隐藏相关字段
+    historyField.style.display = historyChecked ? 'block' : 'none';
+    bookmarksField.style.display = bookmarksChecked ? 'block' : 'none';
+}
+
+function saveCardConfig() {
+    const title = document.getElementById('card-title').value.trim();
+    
+    // 获取选中的数据源
+    const types = [];
+    if (document.getElementById('card-type-bookmarks').checked) {
+        types.push('bookmarks');
+    }
+    if (document.getElementById('card-type-history').checked) {
+        types.push('history');
+    }
+    
+    const filter = document.getElementById('card-filter').value.trim();
+    const folderFilter = document.getElementById('card-folder-filter').value.trim();
+    const historyDays = parseInt(document.getElementById('card-history-days').value);
+    const maxItems = parseInt(document.getElementById('card-max-items').value);
+    
+    if (!title) {
+        alert('请输入卡片标题');
+        return;
+    }
+    
+    if (types.length === 0) {
+        alert('请至少选择一个数据来源');
+        return;
+    }
+    
+    // 验证筛选器语法
+    if (filter) {
+        const validation = FilterParser.validate(filter);
+        if (!validation.valid) {
+            alert(`筛选语法错误: ${validation.error}`);
+            return;
+        }
+    }
+    
+    const cardConfig = {
+        id: currentEditingCardId || generateCardId(),
+        title,
+        types,
+        filter,
+        folderFilter,
+        historyDays,
+        maxItems
+    };
+    
+    if (currentEditingCardId) {
+        // 编辑现有卡片
+        const index = cardsConfig.findIndex(c => c.id === currentEditingCardId);
+        if (index !== -1) {
+            cardsConfig[index] = cardConfig;
+        }
+    } else {
+        // 添加新卡片
+        cardsConfig.push(cardConfig);
+    }
+    
+    saveCardsConfig(cardsConfig);
+    renderAllCards();
+    hideCardConfigModal();
+}
+
+// ====== 布局设置功能 ======
+function showLayoutConfigModal() {
+    document.getElementById('cards-per-row').value = layoutConfig.cardsPerRow;
+    document.getElementById('card-height').value = layoutConfig.cardHeight;
+    document.getElementById('layout-config-modal').style.display = 'block';
+}
+
+function hideLayoutConfigModal() {
+    document.getElementById('layout-config-modal').style.display = 'none';
+}
+
+function saveLayoutConfigFromModal() {
+    const cardsPerRow = parseInt(document.getElementById('cards-per-row').value);
+    const cardHeight = parseInt(document.getElementById('card-height').value);
+    
+    layoutConfig = { cardsPerRow, cardHeight };
+    saveLayoutConfig(layoutConfig);
+    renderAllCards();
+    hideLayoutConfigModal();
+}
+
+// ====== 设置 openInNewTab ======
+export function setOpenInNewTab(value) {
+    openInNewTab = value;
+}
+
+// ====== 初始化 ======
+export function initCardManager() {
+    // 加载配置
+    cardsConfig = loadCardsConfig();
+    layoutConfig = loadLayoutConfig();
+    
+    // 加载数据并渲染卡片
+    loadAllData();
+    
+    // 绑定底部工具栏按钮
+    document.getElementById('add-card-button').addEventListener('click', addCard);
+    document.getElementById('layout-settings-button').addEventListener('click', showLayoutConfigModal);
+    
+    // 绑定卡片配置模态框事件
+    document.getElementById('close-card-config').addEventListener('click', hideCardConfigModal);
+    document.getElementById('cancel-card-config').addEventListener('click', hideCardConfigModal);
+    document.getElementById('save-card-config').addEventListener('click', saveCardConfig);
+    document.getElementById('card-type-bookmarks').addEventListener('change', toggleCardTypeFields);
+    document.getElementById('card-type-history').addEventListener('change', toggleCardTypeFields);
+    
+    // 绑定布局配置模态框事件
+    document.getElementById('close-layout-config').addEventListener('click', hideLayoutConfigModal);
+    document.getElementById('cancel-layout-config').addEventListener('click', hideLayoutConfigModal);
+    document.getElementById('save-layout-config').addEventListener('click', saveLayoutConfigFromModal);
+    
+    // 绑定模态框点击外部关闭
+    document.getElementById('card-config-modal').addEventListener('click', (e) => {
+        if (e.target.id === 'card-config-modal') {
+            hideCardConfigModal();
+        }
+    });
+    document.getElementById('layout-config-modal').addEventListener('click', (e) => {
+        if (e.target.id === 'layout-config-modal') {
+            hideLayoutConfigModal();
+        }
+    });
+}
