--- test-filter.html
@@ -1,165 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-CN">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>筛选器测试</title>
-    <style>
-        body {
-            font-family: Arial, sans-serif;
-            max-width: 800px;
-            margin: 20px auto;
-            padding: 20px;
-        }
-        .test-section {
-            margin: 20px 0;
-            padding: 15px;
-            border: 1px solid #ddd;
-            border-radius: 5px;
-        }
-        input {
-            width: 100%;
-            padding: 8px;
-            margin: 10px 0;
-            font-size: 14px;
-        }
-        .result {
-            margin: 10px 0;
-            padding: 10px;
-            background: #f5f5f5;
-            border-radius: 4px;
-        }
-        .item {
-            padding: 5px;
-            margin: 5px 0;
-            border-left: 3px solid #4285F4;
-            padding-left: 10px;
-        }
-        .match {
-            background: #e8f5e9;
-        }
-        .no-match {
-            background: #ffebee;
-            opacity: 0.5;
-        }
-        h3 {
-            color: #333;
-        }
-        .syntax-hint {
-            font-size: 12px;
-            color: #666;
-            margin: 5px 0;
-        }
-    </style>
-</head>
-<body>
-    <h1>筛选器功能测试</h1>
-    
-    <div class="test-section">
-        <h3>测试数据</h3>
-        <div id="test-data"></div>
-    </div>
-    
-    <div class="test-section">
-        <h3>筛选测试</h3>
-        <div class="syntax-hint">
-            支持语法: 关键词 | title:文本 | url:文本 | /正则/ | title:/正则/ | url:/正则/
-        </div>
-        <input type="text" id="filter-input" placeholder="输入筛选条件，例: github 或 title:新闻 或 /^https/">
-        <button onclick="testFilter()">测试筛选</button>
-        <div id="result"></div>
-    </div>
-
-    <div class="test-section">
-        <h3>预设测试案例</h3>
-        <button onclick="testCase('github')">测试: github</button>
-        <button onclick="testCase('title:新闻')">测试: title:新闻</button>
-        <button onclick="testCase('url:.com')">测试: url:.com</button>
-        <button onclick="testCase('/^https/')">测试: /^https/</button>
-        <button onclick="testCase('title:/^GitHub/')">测试: title:/^GitHub/</button>
-        <button onclick="testCase('url:/github\\.com/')">测试: url:/github\.com/</button>
-    </div>
-
-    <script type="module">
-        import { FilterParser, filterItems } from './src/filterParser.js';
-        
-        // 测试数据
-        const testItems = [
-            { title: 'GitHub - 代码托管平台', url: 'https://github.com' },
-            { title: '今日新闻头条', url: 'https://news.example.com' },
-            { title: 'Google Search', url: 'https://www.google.com' },
-            { title: 'GitHub Actions 教程', url: 'https://docs.github.com/actions' },
-            { title: '技术博客', url: 'http://blog.example.org' },
-            { title: 'Vue.js 中文文档', url: 'https://cn.vuejs.org' },
-            { title: '新闻资讯 - GitHub Trending', url: 'https://github.com/trending' },
-        ];
-
-        // 显示测试数据
-        const testDataDiv = document.getElementById('test-data');
-        testItems.forEach(item => {
-            const itemDiv = document.createElement('div');
-            itemDiv.className = 'item';
-            itemDiv.innerHTML = `<strong>${item.title}</strong><br><small>${item.url}</small>`;
-            testDataDiv.appendChild(itemDiv);
-        });
-
-        // 测试筛选
-        window.testFilter = function() {
-            const filterString = document.getElementById('filter-input').value;
-            runTest(filterString);
-        };
-
-        window.testCase = function(filterString) {
-            document.getElementById('filter-input').value = filterString;
-            runTest(filterString);
-        };
-
-        function runTest(filterString) {
-            const resultDiv = document.getElementById('result');
-            resultDiv.innerHTML = '';
-
-            if (!filterString.trim()) {
-                resultDiv.innerHTML = '<p>请输入筛选条件</p>';
-                return;
-            }
-
-            // 验证语法
-            const validation = FilterParser.validate(filterString);
-            if (!validation.valid) {
-                resultDiv.innerHTML = `<p style="color: red;">语法错误: ${validation.error}</p>`;
-                return;
-            }
-
-            // 创建筛选器
-            const parser = new FilterParser(filterString);
-            const description = parser.getDescription();
-            
-            resultDiv.innerHTML = `<p><strong>筛选规则:</strong> ${description}</p>`;
-
-            // 测试每个项目
-            testItems.forEach(item => {
-                const matches = parser.test(item);
-                const itemDiv = document.createElement('div');
-                itemDiv.className = 'item ' + (matches ? 'match' : 'no-match');
-                itemDiv.innerHTML = `
-                    ${matches ? '✅' : '❌'} 
-                    <strong>${item.title}</strong><br>
-                    <small>${item.url}</small>
-                `;
-                resultDiv.appendChild(itemDiv);
-            });
-
-            // 显示筛选结果统计
-            const filtered = filterItems(filterString, testItems);
-            const statsDiv = document.createElement('div');
-            statsDiv.innerHTML = `<p><strong>匹配结果:</strong> ${filtered.length} / ${testItems.length} 项</p>`;
-            resultDiv.insertBefore(statsDiv, resultDiv.firstChild.nextSibling);
-        }
-
-        // 自动运行一个示例
-        setTimeout(() => {
-            testCase('github');
-        }, 500);
-    </script>
-</body>
-</html>
